Fixed bug clear schedule

+++ ucsdscheduleme/Controllers/HomeController.cs
@@ -116,14 +116,22 @@ namespace ucsdscheduleme.Controllers
             // Call the schedule finding algorithm.
             PossibleSchedules schedules = scheduleRepo.FindScheduleForClasses(courses);
 
-            List<Section> schedule = scheduleRepo.Optimize(optimization, schedules);
+            
 
+            if (!schedules.Any())
+            {
+                // TODO: return error
+                return Json(new ScheduleViewModel());
+            }
+            List<Section> schedule = scheduleRepo.Optimize(optimization, schedules);
+            
             // Get the current user.
             var user = _userManager.GetUserAsync(User).Result;
+
+            // Create user sections to add.
+            var sectionsToAdd = schedule.Select(s => new UserSection { Section = s, User = user });           
             // Get the schedule sections for this user.
             var sectionsToRemove = _context.UserSections.Where(us => us.User == user);
-            // Create user sections to add.
-            var sectionsToAdd = schedule.Select(s => new UserSection { Section = s, User = user });
             // Remove old sections.
             _context.RemoveRange(sectionsToRemove);
             // Add new ones.
